+++ src/lib/comments/CommentForm.svelte
@@ -1,14 +1,19 @@
 <script lang="ts">
-	import type { CommentResponse, Language, LanguageId, MyUserInfo } from 'lemmy-js-client'
+	import type { Language, LanguageId, MyUserInfo } from 'lemmy-js-client'
 	import { createEventDispatcher } from 'svelte'
-	import FlatButton from './buttons/FlatButton.svelte'
-	import Prose from './Prose.svelte'
+	import FlatButton from '../buttons/FlatButton.svelte'
+	import Prose from '../Prose.svelte'
 
-	const dispatch = createEventDispatcher<{ comment: CommentResponse }>()
+	const dispatch = createEventDispatcher<{
+		submit: { content: string; languageId: LanguageId }
+		cancel: undefined
+	}>()
 
 	export let allLanguages: Language[]
+	export let content: string
+	export let disabled: boolean
 	export let myUser: MyUserInfo
-	export let createComment: (content: string, language_id: LanguageId) => Promise<CommentResponse>
+	export let languageId = myUser.discussion_languages[0] ?? 0
 
 	$: myLanguages = myUser.discussion_languages.flatMap(
 		id => allLanguages.find(l => l.id === id) ?? [],
@@ -42,31 +47,18 @@
 	]
 	const placeholder = placeholders[Math.floor(Math.random() * placeholders.length)]
 
-	async function clickReply() {
-		creatingComment = true
-
-		const response = await createComment(text, languageId).catch(e => {
-			if (e instanceof Error) errorMessage = e.message
-			else errorMessage = 'An unexpected error occurred while replying.'
-			console.error('Error while replying:', e)
-		})
-
-		creatingComment = false
-		if (response) dispatch('comment', response)
+	function clickSubmit() {
+		return dispatch('submit', { content: content, languageId })
 	}
 
-	let creatingComment = false
-	let errorMessage = ''
-	let languageId = myUser.discussion_languages[0] ?? 0
 	let previewing = false
-	let text = ''
 </script>
 
 <div class="flex flex-col gap-4">
 	{#if previewing}
 		<div class="rounded-xl bg-surface-container p-4 text-on-surface-container">
-			{#if text}
-				<Prose markdown={text} />
+			{#if content}
+				<Prose markdown={content} />
 			{:else}
 				<p class="text-center text-muted">Nothing to preview</p>
 			{/if}
@@ -76,7 +68,7 @@
 	<textarea
 		{placeholder}
 		class="h-32 w-full resize-y rounded-xl border-0 bg-surface-container p-4 text-on-surface-container focus:border-on-surface-container/25 focus:ring-on-surface-container/25"
-		bind:value={text}
+		bind:value={content}
 	/>
 
 	<!-- Actions -->
@@ -84,7 +76,7 @@
 		<select
 			bind:value={languageId}
 			class="w-40 rounded-md border-none bg-surface-container px-4 py-2 text-on-surface-container"
-			disabled={creatingComment}
+			{disabled}
 		>
 			{#each myLanguages as language (language.id)}
 				<option value={language.id}>{language.name}</option>
@@ -94,31 +86,21 @@
 		<div class="flex flex-row items-center gap-4">
 			<FlatButton
 				class="bg-surface-container px-4 py-2 text-on-surface-container"
+				on:click={() => dispatch('cancel')}
+			>
+				Cancel
+			</FlatButton>
+
+			<FlatButton
+				class="bg-surface-container px-4 py-2 text-on-surface-container"
 				on:click={() => (previewing = !previewing)}
 			>
 				Preview
 			</FlatButton>
 
-			<FlatButton
-				class="bg-surface px-4 py-2 text-on-surface"
-				disabled={creatingComment}
-				on:click={clickReply}
-			>
-				Reply
+			<FlatButton class="bg-surface px-4 py-2 text-on-surface" {disabled} on:click={clickSubmit}>
+				Submit
 			</FlatButton>
 		</div>
 	</div>
-
-	{#if errorMessage}
-		<p
-			class="rounded-md bg-danger-container p-4 text-on-danger-container"
-			on:click={() => (errorMessage = '')}
-			on:keypress={e => {
-				if (e.key === 'Enter') errorMessage = ''
-			}}
-			role="presentation"
-		>
-			{errorMessage}
-		</p>
-	{/if}
 </div>
