+++ src/routes/[site]/u/[user]/+page.svelte
@@ -73,7 +73,7 @@
 				<LimitSelector limit={data.limit ?? 10} />
 			</div>
 
-			{#if data.posts.length}
+			{#if data.posts.length || data.page}
 				<PaginationBar
 					length={data.posts.length}
 					limit={data.limit ?? 10}
@@ -90,7 +90,7 @@
 				showCommunity={true}
 				site={data.site_view.site}
 			/>
-			{#if data.posts.length}
+			{#if data.posts.length || data.page}
 				<PaginationBar
 					length={data.posts.length}
 					limit={data.limit ?? 10}
